refactor(teststack): split features.go into per-resource files for ownership
Move each feature's setup function into its own file mirroring the
internal/resources/<name>/ pattern:

  asserts.go              installAsserts + assertsErr
  cloudintegrations.go    installCloudIntegrations (no-op stub)
  fleetmanagement.go      installFleet
  k6.go                   installK6
  oncall.go               getOnCallURL
  syntheticmonitoring.go  installSM + smAPIURL

features.go keeps only the feature dispatcher (constants + parseFeatures).
waitStackHealthy moves to stack.go where the other stack-readiness
helpers live. TestSMAPIURL moves into syntheticmonitoring_test.go to
match the production file split.

Add CODEOWNERS.in entries so each per-feature file is owned by the
same Grafana team that owns the corresponding internal/resources/<name>/
package — teams can now self-service changes to how their feature is
provisioned on ephemeral CI stacks without going through platform
monitoring.
