+++ tools/teststack/features.go
@@ -1,20 +1,20 @@
 package main
 
 import (
-	"context"
-	"encoding/json"
 	"fmt"
-	"net/http"
-	"net/url"
 	"strings"
-	"time"
-
-	assertsapi "github.com/grafana/grafana-asserts-public-clients/go/gcom"
-	"github.com/grafana/grafana-com-public-clients/go/gcom"
-	smapi "github.com/grafana/synthetic-monitoring-api-go-client"
 )
 
-// known feature identifiers accepted by --features.
+// Known feature identifiers accepted by --features. Each non-basic feature
+// has its own file in this package mirroring the corresponding
+// internal/resources/<name>/ package, so each Grafana team can own the
+// teststack setup for their own product via .github/CODEOWNERS.
+//
+// To add a new feature:
+//  1. Add a const below and include it in the parseFeatures switch.
+//  2. Create tools/teststack/<feature>.go with the install/setup function.
+//  3. Wire the install function into up.go.
+//  4. Add a CODEOWNERS entry in .github/CODEOWNERS.in.
 const (
 	featureBasic        = "basic"
 	featureK6           = "k6"
@@ -44,262 +44,4 @@ func parseFeatures(spec string) (map[string]bool, error) {
 		}
 	}
 	return out, nil
-}
-
-// installK6 calls the k6 install endpoint and returns the K6 access token and
-// API URL. Mirrors the logic in resource_k6_installation.go.
-func installK6(ctx context.Context, capToken string, info *stackInfo, cloudAPIBase string) (token, apiURL string, err error) {
-	apiURL = "https://api.k6.io"
-	url := fmt.Sprintf("%s/v3/account/grafana-app/start", apiURL)
-
-	req, err := http.NewRequestWithContext(ctx, http.MethodPut, url, nil)
-	if err != nil {
-		return "", "", err
-	}
-	req.Header.Set("X-Stack-Id", fmt.Sprintf("%d", info.ID))
-	req.Header.Set("X-Grafana-Key", capToken)
-	req.Header.Set("X-Grafana-Service-Token", info.AdminSAToken)
-	req.Header.Set("X-Grafana-User", "admin")
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return "", "", fmt.Errorf("k6 install request: %w", err)
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode >= 300 {
-		return "", "", fmt.Errorf("k6 install returned status %d", resp.StatusCode)
-	}
-
-	var out struct {
-		V3GrafanaToken string `json:"v3_grafana_token"`
-		OrganizationID string `json:"organization_id"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&out); err != nil {
-		return "", "", fmt.Errorf("decode k6 install response: %w", err)
-	}
-	if out.V3GrafanaToken == "" {
-		return "", "", fmt.Errorf("k6 install response missing v3_grafana_token")
-	}
-	_ = cloudAPIBase
-	return out.V3GrafanaToken, apiURL, nil
-}
-
-// smAPIURL returns the Synthetic Monitoring API URL for a given region slug,
-// mirroring the table in resource_synthetic_monitoring_installation.go.
-func smAPIURL(regionSlug string) string {
-	exceptions := map[string]string{
-		"au":              "https://synthetic-monitoring-api-au-southeast.grafana.net",
-		"eu":              "https://synthetic-monitoring-api-eu-west.grafana.net",
-		"prod-gb-south-0": "https://synthetic-monitoring-api-gb-south.grafana.net",
-		"us":              "https://synthetic-monitoring-api.grafana.net",
-		"us-azure":        "https://synthetic-monitoring-api-us-central-7.grafana.net",
-	}
-	if v, ok := exceptions[regionSlug]; ok {
-		return v
-	}
-	return fmt.Sprintf("https://synthetic-monitoring-api-%s.grafana.net", strings.TrimPrefix(regionSlug, "prod-"))
-}
-
-// installSM installs Synthetic Monitoring on a stack and returns the SM access
-// token plus the SM API URL.
-func installSM(ctx context.Context, capToken string, info *stackInfo) (token, apiURL string, err error) {
-	apiURL = smAPIURL(info.RegionSlug)
-	client := smapi.NewClient(apiURL, "", nil)
-	client.SetCustomClientID("teststack")
-	client.SetCustomClientVersion("0.1")
-
-	resp, err := client.Install(ctx, info.ID, info.HmInstancePromID, info.HlInstanceID, capToken)
-	if err != nil {
-		return "", "", fmt.Errorf("sm install: %w", err)
-	}
-	return resp.AccessToken, apiURL, nil
-}
-
-// installFleet returns basic-auth credentials and a URL for Grafana Fleet
-// Management. Fleet management is auto-provisioned with a stack; the auth
-// format per the provider docs is `{fleet_management_user_id}:{cap_token}`
-// (see docs/index.md). The user ID comes from the gcom "agent management"
-// fields (legacy name for the same backend); the password is the org-level
-// Cloud Access Policy token that authenticates the caller.
-func installFleet(ctx context.Context, client *gcom.APIClient, capToken string, info *stackInfo) (auth, apiURL string, err error) {
-	stack, _, err := client.InstancesAPI.GetInstance(ctx, info.Slug).Execute()
-	if err != nil {
-		return "", "", fmt.Errorf("get instance for fleet management URL: %w", gcomErr(err))
-	}
-
-	apiURL = strings.TrimSpace(stack.AgentManagementInstanceUrl)
-	if apiURL == "" {
-		return "", "", fmt.Errorf("stack %q has no fleet management URL configured", info.Slug)
-	}
-
-	fleetUserID := stack.AgentManagementInstanceId
-	if fleetUserID == 0 {
-		return "", "", fmt.Errorf("stack %q has no fleet management user id configured", info.Slug)
-	}
-
-	auth = fmt.Sprintf("%d:%s", int64(fleetUserID), capToken)
-	return auth, apiURL, nil
-}
-
-// installAsserts performs the Asserts onboarding flow on a freshly-created
-// stack. Without this, Asserts resource tests hit 403 Forbidden because the
-// stack is in the `not_initialized` state. The flow mirrors what
-// resource_stack.go does:
-//
-//  1. PUT /v2/stack — provision tokens
-//  2. PUT /v2/stack/dataset — manually configure a 'prometheus' dataset
-//  3. POST /v2/stack/enable — enable the stack
-//
-// We configure the prometheus dataset explicitly rather than relying on
-// auto-setup because a freshly-created Grafana Cloud stack does not yet
-// have any provisioned datasources visible to auto-detection, so the
-// auto-setup endpoint finds nothing and the subsequent enable step
-// fails the sanity-check with 422. A 'prometheus' dataset is always
-// present on a Grafana Cloud stack (it's the stack's own Mimir tenant).
-//
-// All three calls go through the Grafana plugin proxy on the stack itself,
-// authenticated with the per-stack Admin SA token.
-func installAsserts(ctx context.Context, capToken string, info *stackInfo) error {
-	cfg := assertsapi.NewConfiguration()
-	u, err := url.Parse(info.URL)
-	if err != nil {
-		return fmt.Errorf("parse stack URL %q: %w", info.URL, err)
-	}
-	cfg.Host = u.Host
-	cfg.Scheme = u.Scheme
-	cfg.Servers = assertsapi.ServerConfigurations{
-		{
-			URL: fmt.Sprintf("%s://%s/api/plugins/grafana-asserts-app/resources/asserts/api-server", u.Scheme, u.Host),
-		},
-	}
-	cfg.HTTPClient = &http.Client{Timeout: 60 * time.Second}
-	if cfg.DefaultHeader == nil {
-		cfg.DefaultHeader = make(map[string]string)
-	}
-	cfg.DefaultHeader["Authorization"] = "Bearer " + info.AdminSAToken
-	cfg.DefaultHeader["Content-Type"] = "application/json"
-	client := assertsapi.NewAPIClient(cfg)
-
-	stackIDStr := fmt.Sprintf("%d", info.ID)
-
-	// Step 1: provision tokens. Reuse the org-level CAP token for gcom /
-	// mimir / detector, and the stack-scoped Admin SA token for the Grafana
-	// token. This matches what resource_stack.go does when a user calls
-	// `grafana_asserts_stack` with the same inputs.
-	stackDto := assertsapi.NewStackDto()
-	stackDto.SetGcomToken(capToken)
-	stackDto.SetMimirToken(capToken)
-	stackDto.SetAssertionDetectorToken(capToken)
-	stackDto.SetGrafanaToken(info.AdminSAToken)
-	if _, err := client.StackControllerAPI.PutV2Stack(ctx).
-		StackDto(*stackDto).
-		XScopeOrgID(stackIDStr).
-		Execute(); err != nil {
-		return fmt.Errorf("asserts PUT /v2/stack: %w", assertsErr(err))
-	}
-
-	// Step 2: configure a 'prometheus' dataset. Asserts ships with
-	// 'kubernetes', 'otel', 'prometheus', and 'aws' as valid dataset
-	// types; 'prometheus' is the only one guaranteed to exist on a
-	// freshly-created Grafana Cloud stack (the stack's own Mimir).
-	//
-	// The dataset endpoint requires at least one filter group ("groups
-	// Required" from the API). We provide a minimal filter that uses the
-	// conventional Prometheus `environment` label as the env label, with
-	// a single env name of `prod`. This satisfies the API validator
-	// without filtering any real metrics out — the metrics-generator
-	// emits this label automatically on most pipelines and the test
-	// suite doesn't depend on a specific environment value.
-	datasetDto := assertsapi.NewStackDatasetDto("prometheus")
-	filterGroup := *assertsapi.NewStackFilterGroupDto()
-	filterGroup.SetEnvLabel("environment")
-	filterGroup.SetEnvName("prod")
-	datasetDto.FilterGroups = []assertsapi.StackFilterGroupDto{filterGroup}
-	if _, _, err := client.StackControllerAPI.UpdateDataset(ctx).
-		StackDatasetDto(*datasetDto).
-		XScopeOrgID(stackIDStr).
-		Execute(); err != nil {
-		return fmt.Errorf("asserts PUT /v2/stack/dataset (prometheus): %w", assertsErr(err))
-	}
-
-	// Step 3: enable. With the prometheus dataset configured the sanity
-	// check passes and the stack flips to enabled=true.
-	if _, _, err := client.StackControllerAPI.EnableV2Stack(ctx).
-		XScopeOrgID(stackIDStr).
-		Execute(); err != nil {
-		return fmt.Errorf("asserts POST /v2/stack/enable: %w", assertsErr(err))
-	}
-
-	return nil
-}
-
-// assertsErr enriches an asserts API error with the response body, which is
-// where the actual 422/409 detail lives. Mirrors gcomErr for the assertsapi
-// generated client.
-func assertsErr(err error) error {
-	if err == nil {
-		return nil
-	}
-	type bodyError interface {
-		Body() []byte
-	}
-	if be, ok := err.(bodyError); ok && len(be.Body()) > 0 {
-		return fmt.Errorf("%w: %s", err, strings.TrimSpace(string(be.Body())))
-	}
-	return err
-}
-
-// getOnCallURL returns the per-stack OnCall API URL from gcom. OnCall is
-// a regional service: the URL host varies by region (e.g.
-// https://oncall-prod-eu-west-0.grafana.net), so we can't hard-code it.
-// The provider requires `oncall_url` to be set before it builds the OnCall
-// client, so the caller exports this value as GRAFANA_ONCALL_URL.
-func getOnCallURL(ctx context.Context, client *gcom.APIClient, info *stackInfo) (string, error) {
-	conn, _, err := client.InstancesAPI.GetConnections(ctx, info.Slug).Execute()
-	if err != nil {
-		return "", fmt.Errorf("get instance connections for OnCall URL: %w", gcomErr(err))
-	}
-	if conn.OncallApiUrl.IsSet() {
-		if v := conn.OncallApiUrl.Get(); v != nil && *v != "" {
-			return *v, nil
-		}
-	}
-	return "", fmt.Errorf("stack %q has no OnCall API URL configured", info.Slug)
-}
-
-// installCloudIntegrations is a no-op stub: cloud integrations are enabled
-// per-stack via the easystart plugin. Installation happens transparently when
-// the tests call the integrations API, so no upfront provisioning is needed.
-func installCloudIntegrations(ctx context.Context, info *stackInfo) error {
-	_ = ctx
-	_ = info
-	return nil
-}
-
-// waitStackHealthy performs a lightweight GET / on the stack URL to confirm
-// the Grafana microservices are up. New stacks frequently return 503/504 for
-// tens of seconds after the gcom status flips to "active", so we poll until
-// the stack actually responds.
-func waitStackHealthy(ctx context.Context, info *stackInfo, timeout time.Duration) error {
-	healthCtx, cancel := context.WithTimeout(ctx, timeout)
-	defer cancel()
-
-	url := strings.TrimSuffix(info.URL, "/") + "/api/health"
-	httpClient := &http.Client{Timeout: 10 * time.Second}
-	return pollUntil(healthCtx, 5*time.Second, func(c context.Context) (bool, error) {
-		req, err := http.NewRequestWithContext(c, http.MethodGet, url, nil)
-		if err != nil {
-			return false, err
-		}
-		resp, err := httpClient.Do(req)
-		if err != nil {
-			return false, err
-		}
-		_ = resp.Body.Close()
-		if resp.StatusCode >= 200 && resp.StatusCode < 400 {
-			return true, nil
-		}
-		return false, fmt.Errorf("stack health status %d", resp.StatusCode)
-	})
 }
